requirements.txt
# ============================================
# DEPENDENCIAS DEL PROYECTO
# Sistema de Gestión Educativa - Colegio Miguel Grau
# ============================================

# Framework Web
Flask==3.0.0
Werkzeug==3.0.1

# Base de Datos
Flask-MySQLdb==1.0.1
mysqlclient==2.2.4

# CORS y Seguridad
Flask-CORS==4.0.0
PyJWT==2.8.0

# Utilidades
python-dotenv==1.0.0

# Servidor WSGI (opcional para producción)
gunicorn==21.2.0
